+++ processor/src/test/resources/expected/com/example/prop/React4j_ObservableProp.java
@@ -6,6 +6,7 @@ import arez.ObservableValue;
 import arez.Observer;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
+import arez.annotations.DepType;
 import arez.annotations.Executor;
 import arez.annotations.Feature;
 import arez.annotations.InjectMode;
@@ -110,6 +111,7 @@ abstract class React4j_ObservableProp extends ObservableProp {
       name = "render",
       priority = Priority.LOW,
       executor = Executor.EXTERNAL,
+      depType = DepType.AREZ_OR_NONE,
       observeLowerPriorityDependencies = true,
       reportResult = false
   )